+++ docs/conf.py
@@ -41,12 +41,12 @@ extensions = [
     "sphinx_js",
     "sphinx_click",
     "autodocsumm",
-    "sphinx_panels",
     "sphinx_pyodide",
     "sphinx_argparse_cli",
     "versionwarning.extension",
     "sphinx_issues",
     "sphinx_autodoc_typehints",
+    "sphinx_design",  # Used for tabs in building-from-sources.md
 ]
 
 
@@ -73,10 +73,11 @@ autodoc_default_flags = ["members", "inherited-members"]
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3.10", None),
     "micropip": (f"https://micropip.pyodide.org/en/v{micropip.__version__}/", None),
+    "numpy": ("https://numpy.org/doc/stable/", None),
 }
 
 # Add modules to be mocked.
-mock_modules = ["ruamel.yaml", "tomli"]
+mock_modules = ["tomli"]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
@@ -184,14 +185,45 @@ if IN_READTHEDOCS:
     console_path.write_text("".join(console_html))
 
 
+def docs_argspec(argspec: str) -> Any:
+    """Decorator to override the argument spec of the function that is
+    rendered in the docs.
+
+    If the documentation finds a __wrapped__ attribute, it will use that to
+    render the argspec instead of the function itself. Assign an appropriate
+    fake function to __wrapped__ with the desired argspec.
+    """
+
+    def dec(func):
+        d = func.__globals__
+        TEMP_NAME = "_xxxffff___"
+        assert TEMP_NAME not in d
+        code = dedent(
+            f"""\
+            def {TEMP_NAME}{argspec}:
+                pass
+            """
+        )
+        # exec in func.__globals__ context so that type hints don't case NameErrors.
+        exec(code, d)
+        f = d.pop(TEMP_NAME)
+
+        # # Run update_wrapper but keep f's annotations and don't set
+        # # f.__wrapped__ (would cause an infinite loop!)
+        annotations = f.__annotations__
+        update_wrapper(f, func)
+        f.__annotations__ = annotations
+        del f.__wrapped__
+
+        # Set wrapper to be our fake function
+        func.__wrapped__ = f
+
+        return func
+
+    return dec
+
+
 if IN_SPHINX:
-    # Compatibility shims. sphinx-js and sphinxcontrib-napoleon have not been updated for Python 3.10
-    import collections
-    from typing import Callable, Mapping
-
-    collections.Mapping = Mapping  # type: ignore[attr-defined]
-    collections.Callable = Callable  # type: ignore[attr-defined]
-
     base_dir = Path(__file__).resolve().parent.parent
     path_dirs = [
         str(base_dir),
@@ -204,9 +236,16 @@ if IN_SPHINX:
 
     from sphinx.domains.javascript import JavaScriptDomain, JSXRefRole
 
-    JavaScriptDomain.roles["func"] = JSXRefRole()
+    JavaScriptDomain.roles["class"] = JSXRefRole()
 
-    import micropip  # noqa: F401
+    import builtins
+    from functools import update_wrapper
+    from textwrap import dedent
+
+    # override docs_argspec, _pyodide.docs_argspec will read this value back.
+    # Must do this before importing pyodide!
+    setattr(builtins, "--docs_argspec--", docs_argspec)
+
     import pyodide
     from pyodide.ffi import JsProxy
 
@@ -257,14 +296,12 @@ def globalReplace(app, docname, source):
 
 global_replacements = {"{{PYODIDE_CDN_URL}}": CDN_URL}
 
+always_document_param_types = True
+
 
 def typehints_formatter(annotation, config):
-    """Adjust the rendering of Literal types.
-
-    The literal values should be ``rendered as code``.
-    """
+    """Adjust the rendering of various types that sphinx_autodoc_typehints mishandles"""
     from sphinx_autodoc_typehints import (
-        get_annotation_args,
         get_annotation_class_name,
         get_annotation_module,
     )
@@ -272,15 +309,15 @@ def typehints_formatter(annotation, config):
     try:
         module = get_annotation_module(annotation)
         class_name = get_annotation_class_name(annotation, module)
-        args = get_annotation_args(annotation, module, class_name)
     except ValueError:
         return None
     full_name = f"{module}.{class_name}"
-    if full_name == "typing.Literal":
-        formatted_args = "\\[{}]".format(
-            ", ".join("``{}``".format(repr(arg)) for arg in args)
-        )
-        return f":py:data:`~{full_name}`{formatted_args}"
+    if full_name == "typing.TypeVar":
+        # The way sphinx-autodoc-typehints renders TypeVar is too noisy for my
+        # taste
+        return f"``{annotation.__name__}``"
+    if full_name == "ast.Module":
+        return "`Module <https://docs.python.org/3/library/ast.html#module-ast>`_"
     return None