+++ docs/sphinx_pyodide/tests/test_directives.py
@@ -1,17 +1,10 @@
-import collections
 import gzip
 import json
 import sys
-import types
 from pathlib import Path
-from typing import Mapping, Union
 
 from docutils.frontend import OptionParser
 from docutils.utils import new_document
-
-# Shim sphinx-js Python 3.10 compatibility
-collections.Mapping = Mapping  # type: ignore[attr-defined]
-types.Union = Union  # type: ignore[attr-defined]
 from sphinx_js.suffix_tree import SuffixTree
 from sphinx_js.typedoc import Analyzer as TsAnalyzer
 
@@ -214,7 +207,7 @@ def test_summary():
 
 def test_type_name():
     tn = inner_analyzer._type_name
-    assert tn({"name": "void", "type": "intrinsic"}) == "void"
+    assert tn({"name": "void", "type": "intrinsic"}) == ":js:data:`void`"
     assert tn({"value": None, "type": "literal"}) == "null"
     assert (
         tn(
@@ -223,8 +216,8 @@ def test_type_name():
                 "type": "reference",
                 "typeArguments": [{"name": "string", "type": "intrinsic"}],
             }
-        )
-        == "Promise<string>"
+        ).strip()
+        == r":js:data:`Promise`\ **<**\ :js:data:`string`\ **>**"
     )
 
     assert (
@@ -235,8 +228,8 @@ def test_type_name():
                 "targetType": {"name": "PyProxy", "type": "reference"},
                 "type": "predicate",
             }
-        )
-        == "boolean (typeguard for PyProxy)"
+        ).strip()
+        == ":js:data:`boolean` (typeguard for :js:class:`PyProxy`)"
     )
 
     assert (
@@ -263,8 +256,8 @@ def test_type_name():
                 },
                 "type": "reflection",
             }
-        )
-        == "(message: string) => void"
+        ).strip()
+        == r"\ **(**\ \ **message:** :js:data:`string`\ **) =>** :js:data:`void`"
     )
 
     assert (
@@ -292,8 +285,8 @@ def test_type_name():
                     }
                 ],
             }
-        )
-        == "Iterable<[key: string, value: any]>"
+        ).strip()
+        == r":js:data:`Iterable`\ **<**\ \ **[**\ \ **key:** :js:data:`string`\ **,** \ **value:** :js:data:`any`\ **]** \ **>**"
     )
 
     assert (
@@ -317,8 +310,8 @@ def test_type_name():
                 },
                 "type": "reflection",
             }
-        )
-        == "{[key: string]: string}"
+        ).strip()
+        == r"""\ **{**\ \ **[key:** :js:data:`string`\ **]:** :js:data:`string`\ **}**\ """.strip()
     )
 
     assert (
@@ -350,6 +343,6 @@ def test_type_name():
                 },
                 "type": "reflection",
             }
-        )
-        == "{cache: PyProxyCache, destroyed_msg?: string, ptr: number}"
+        ).strip()
+        == r"""\ **{**\ \ **cache:** :js:class:`PyProxyCache`\ **,** \ **destroyed_msg?:** :js:data:`string`\ **,** \ **ptr:** :js:data:`number`\ **}**\ """.strip()
     )