+++ pyodide-build/pyodide_build/cli/create_zipfile.py
@@ -20,7 +20,7 @@ def main(
     Bundle Python standard libraries into a zip file.
     """
     create_zipfile(libdir, output, pycompile=pycompile, filterfunc=None)
-    typer.echo(f"Zip file created at {output}")
+    typer.echo(f"Zip file created at {output.resolve()}")
 
 
 main.typer_kwargs = {"hidden": True}  # type: ignore[attr-defined]