+++ src/js/pyodide.ts
@@ -367,7 +367,13 @@ If you updated the Pyodide version, make sure you also updated the 'indexURL' pa
 
   const pyodide_py_tar = await pyodide_py_tar_promise;
   unpackPyodidePy(Module, pyodide_py_tar);
-  API.rawRun("import _pyodide_core");
+  let [err, captured_stderr] = API.rawRun("import _pyodide_core");
+  if (err) {
+    Module.API.fatal_loading_error(
+      "Failed to import _pyodide_core\n",
+      captured_stderr,
+    );
+  }
 
   const pyodide = finalizeBootstrap(API, config);
 
@@ -380,7 +386,7 @@ If you updated the Pyodide version, make sure you also updated the 'indexURL' pa
   await API.packageIndexReady;
 
   let importhook = API._pyodide._importhook;
-  importhook.register_module_not_found_hook(API.repodata_packages);
+  importhook.register_module_not_found_hook(API._import_name_to_package_name);
 
   if (API.repodata_info.version !== version) {
     throw new Error("Lock file version doesn't match Pyodide version");