+++ src/py/_pyodide/_core_docs.py
@@ -15,6 +15,8 @@ from functools import reduce
 from types import TracebackType
 from typing import IO, Any, Awaitable, Generic, TypeVar, overload
 
+from .docs_argspec import docs_argspec
+
 # All docstrings for public `core` APIs should be extracted from here. We use
 # the utilities in `docstring.py` and `docstring.c` to format them
 # appropriately.
@@ -107,14 +109,14 @@ class JsProxy(metaclass=_JsProxyMetaClass):
         """An id number which can be used as a dictionary/set key if you want to
         key on JavaScript object identity.
 
-        If two `JsProxy` are made with the same backing JavaScript object, they
-        will have the same `js_id`.
+        If two ``JsProxy`` are made with the same backing JavaScript object, they
+        will have the same ``js_id``.
         """
         return 0
 
     @property
     def typeof(self) -> str:
-        """Returns the JavaScript type of the JsProxy.
+        """Returns the JavaScript type of the ``JsProxy``.
 
         Corresponds to `typeof obj;` in JavaScript. You may also be interested
         in the `constuctor` attribute which returns the type as an object.
@@ -169,8 +171,8 @@ class JsProxy(metaclass=_JsProxyMetaClass):
         Parameters
         ----------
         depth:
-            You can use this argument to limit the depth of the conversion. If a
-            shallow conversion is desired, set `depth` to 1.
+            Limit the depth of the conversion. If a shallow conversion is
+            desired, set ``depth`` to 1.
 
         default_converter:
 
@@ -291,7 +293,7 @@ class JsPromise(JsProxy):
         """
         raise NotImplementedError
 
-    def finally_(self, onfinally: Callable[[Any], Any], /) -> "JsPromise":
+    def finally_(self, onfinally: Callable[[], Any], /) -> "JsPromise":
         """The ``Promise.finally`` API, wrapped to manage the lifetimes of
         the handler.
 
@@ -324,7 +326,8 @@ class JsBuffer(JsProxy):
     def to_memoryview(self) -> memoryview:
         """Convert a buffer to a memoryview.
 
-        Copies the data once. This currently has the same effect as :any:`to_py`.
+        Copies the data once. This currently has the same effect as
+        :py:meth:`~JsArray.to_py`.
         """
         raise NotImplementedError
 
@@ -410,8 +413,7 @@ class JsBuffer(JsProxy):
 
         Copies the data twice.
 
-        The encoding argument will be passed to the Javascript `TextDecoder
-        <(https://developer.mozilla.org/en-US/docs/Web/API/TextDecoder)>`_
+        The encoding argument will be passed to the :js:class:`TextDecoder`
         constructor. It should be one of the encodings listed in `the table here
         <https://encoding.spec.whatwg.org/#names-and-labels>`_. The default
         encoding is utf8.
@@ -420,7 +422,7 @@ class JsBuffer(JsProxy):
 
 
 class JsArray(JsProxy, Generic[T]):
-    """A JsProxy of an array, node list, or typed array"""
+    """A JsProxy of an :js:class:`Array`, :js:class:`NodeList`, or :js:class:`TypedArray`"""
 
     _js_type_flags = ["IS_ARRAY", "IS_NODE_LIST", "IS_TYPEDARRAY"]
 
@@ -440,13 +442,13 @@ class JsArray(JsProxy, Generic[T]):
         """Extend array by appending elements from the iterable."""
 
     def __reversed__(self) -> Iterator[T]:
-        """Return a reverse iterator over the Array."""
+        """Return a reverse iterator over the :js:class:`Array`."""
         raise NotImplementedError
 
     def pop(self, /, index: int = -1) -> T:
-        """Remove and return item at index (default last).
+        """Remove and return the ``item`` at ``index`` (default last).
 
-        Raises IndexError if list is empty or index is out of range.
+        Raises :any:`IndexError` if list is empty or index is out of range.
         """
         raise NotImplementedError
 
@@ -457,9 +459,9 @@ class JsArray(JsProxy, Generic[T]):
         """Append object to the end of the list."""
 
     def index(self, /, value: T, start: int = 0, stop: int = sys.maxsize) -> int:
-        """Return first index of value.
+        """Return first ``index`` at which ``value`` appears in the ``Array``.
 
-        Raises ValueError if the value is not present.
+        Raises :any:`ValueError` if the value is not present.
         """
         raise NotImplementedError
 
@@ -502,9 +504,9 @@ class JsTypedArray(JsBuffer, JsArray[int]):
 class JsMap(JsProxy, Generic[KT, VTco]):
     """A JavaScript Map
 
-    To be considered a map, a JavaScript object must have a ``.get`` method, it
-    must have a ``.size`` or a ``.length`` property which is a number
-    (idiomatically it should be called ``.size``) and it must be iterable.
+    To be considered a map, a JavaScript object must have a ``get`` method, it
+    must have a ``size`` or a ``length`` property which is a number
+    (idiomatically it should be called ``size``) and it must be iterable.
     """
 
     _js_type_flags = ["HAS_GET | HAS_LENGTH | IS_ITERABLE", "IS_OBJECT_MAP"]
@@ -522,27 +524,28 @@ class JsMap(JsProxy, Generic[KT, VTco]):
         raise NotImplementedError
 
     def keys(self) -> KeysView[KT]:
-        """Return a :any:`KeysView` for the map."""
+        """Return a :py:class:`~collections.abc.KeysView` for the map."""
         raise NotImplementedError
 
     def items(self) -> ItemsView[KT, VTco]:
-        """Return a :any:`ItemsView` for the map."""
+        """Return a :py:class:`~collections.abc.ItemsView` for the map."""
         raise NotImplementedError
 
     def values(self) -> ValuesView[VTco]:
-        """Return a :any:`ValuesView` for the map."""
+        """Return a :py:class:`~collections.abc.ValuesView` for the map."""
         raise NotImplementedError
 
     @overload
-    def get(self, key: KT) -> VTco | None:
+    def get(self, key: KT, /) -> VTco | None:
         ...
 
     @overload
-    def get(self, key: KT, default: VTco | T) -> VTco | T:
+    def get(self, key: KT, default: VTco | T, /) -> VTco | T:
         ...
 
-    def get(self, key, default=None):
-        """If key in self, returns self[key]. Otherwise returns default."""
+    @docs_argspec("(self, key: KT, default: VTco | None, /) -> VTco")
+    def get(self, key: KT, default: Any = None, /) -> VTco:
+        r"""If ``key in self``, returns ``self[key]``. Otherwise returns ``default``."""
         raise NotImplementedError
 
 
@@ -550,10 +553,9 @@ class JsMap(JsProxy, Generic[KT, VTco]):
 class JsMutableMap(JsMap[KT, VT], Generic[KT, VT]):
     """A JavaScript mutable map
 
-    To be considered a mutable map, a JavaScript object must have a ``.get``
-    method, a ``.has`` method, a ``.size`` or a ``.length`` property which is a
-    number (idiomatically it should be called ``.size``) and it must be
-    iterable.
+    To be considered a mutable map, a JavaScript object must have a ``get``
+    method, a ``has`` method, a ``size`` or a ``length`` property which is a
+    number (idiomatically it should be called ``size``) and it must be iterable.
 
     Instances of the JavaScript builtin ``Map`` class are ``JsMutableMap`` s.
     Also proxies returned by :any:`JsProxy.as_object_map` are instances of
@@ -563,28 +565,29 @@ class JsMutableMap(JsMap[KT, VT], Generic[KT, VT]):
     _js_type_flags = ["HAS_GET | HAS_SET | HAS_LENGTH | IS_ITERABLE", "IS_OBJECT_MAP"]
 
     @overload
-    def pop(self, __key: KT) -> VT:
+    def pop(self, key: KT, /) -> VT:
         ...
 
     @overload
-    def pop(self, __key: KT, __default: VT | T = ...) -> VT | T:
+    def pop(self, key: KT, default: VT | T = ..., /) -> VT | T:
         ...
 
-    def pop(self, key, default=None):
-        """If key in self, return self[key] and remove key from self. Otherwise
-        returns default.
+    @docs_argspec("(self, key: KT, default: VT | None = None, /) -> VT")
+    def pop(self, key: KT, default: Any = None, /) -> Any:
+        r"""If ``key in self``, return ``self[key]`` and remove key from ``self``. Otherwise
+        returns ``default``.
         """
         raise NotImplementedError
 
     def setdefault(self, key: KT, default: VT | None = None) -> VT:
-        """If key in self, return self[key]. Otherwise
-        sets self[key] = default and returns default.
+        """If ``key in self``, return ``self[key]``. Otherwise
+        sets ``self[key] = default`` and returns ``default``.
         """
         raise NotImplementedError
 
-    def popitem(self) -> tuple[KT, KT]:
-        """Remove some arbitrary key, value pair from the map and returns the
-        (key, value) tuple.
+    def popitem(self) -> tuple[KT, VT]:
+        """Remove some arbitrary ``key, value`` pair from the map and returns the
+        ``(key, value)`` tuple.
         """
         raise NotImplementedError
 
@@ -603,10 +606,25 @@ class JsMutableMap(JsMap[KT, VT], Generic[KT, VT]):
     def update(self, **kwargs: VT) -> None:
         ...
 
-    def update(self, other, **kwargs):
-        """Updates self from other and kwargs.
+    @docs_argspec(
+        "(self, other : Mapping[KT, VT] | Iterable[tuple[KT, VT]] = None , /, **kwargs) -> None"
+    )
+    def update(self, *args: Any, **kwargs: Any) -> None:
+        r"""Updates ``self`` from ``other`` and ``kwargs``.
 
-        If ``other`` is present and is a :any:`Mapping` or has a ``keys`` method, does
+        Parameters
+        ----------
+            other:
+
+                Either a mapping or an iterable of pairs. This can be left out.
+
+            kwargs:  ``VT``
+
+                Extra key-values pairs to insert into the map. Only usable for
+                inserting extra strings.
+
+        If ``other`` is present and is a :any:`Mapping` or has a ``keys``
+        method, does
 
         .. code-block:: python
 
@@ -639,8 +657,8 @@ class JsMutableMap(JsMap[KT, VT], Generic[KT, VT]):
 class JsIterator(JsProxy, Generic[Tco]):
     """A JsProxy of a JavaScript iterator.
 
-    An object is a JsIterator if it has a `next` method and either has a
-    Symbol.iterator or has no Symbol.asyncIterator.
+    An object is a :any:`JsAsyncIterator` if it has a :js:meth:`~Iterator.next` method and either has a
+    :js:data:`Symbol.iterator` or has no :js:data:`Symbol.asyncIterator`.
     """
 
     _js_type_flags = ["IS_ITERATOR"]
@@ -655,8 +673,9 @@ class JsIterator(JsProxy, Generic[Tco]):
 class JsAsyncIterator(JsProxy, Generic[Tco]):
     """A JsProxy of a JavaScript async iterator.
 
-    An object is a JsAsyncIterator if it has a `next` method and either has a
-    Symbol.asyncIterator or has no Symbol.iterator.
+    An object is a :any:`JsAsyncIterator` if it has a
+    :js:meth:`~AsyncIterator.next` method and either has a
+    :js:data:`Symbol.asyncIterator` or has no :js:data:`Symbol.iterator`
     """
 
     _js_type_flags = ["IS_ASYNC_ITERATOR"]
@@ -671,7 +690,7 @@ class JsAsyncIterator(JsProxy, Generic[Tco]):
 class JsIterable(JsProxy, Generic[Tco]):
     """A JavaScript iterable object
 
-    A JavaScript object is iterable if it has a ``Symbol.iterator`` method.
+    A JavaScript object is iterable if it has a :js:data:`Symbol.iterator` method.
     """
 
     _js_type_flags = ["IS_ITERABLE"]
@@ -683,7 +702,8 @@ class JsIterable(JsProxy, Generic[Tco]):
 class JsAsyncIterable(JsProxy, Generic[Tco]):
     """A JavaScript async iterable object
 
-    A JavaScript object is async iterable if it has a ``Symbol.asyncIterator`` method.
+    A JavaScript object is async iterable if it has a :js:data:`Symbol.asyncIterator`
+    method.
     """
 
     _js_type_flags = ["IS_ASYNC_ITERABLE"]
@@ -696,14 +716,11 @@ class JsGenerator(JsIterable[Tco], Generic[Tco, Tcontra, Vco]):
     """A JavaScript generator
 
     A JavaScript object is treated as a generator if its
-    `Symbol.toStringTag <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol/toStringTag>`_
-    is ``"Generator"``. Most likely this will be because it is a true
-    `generator <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Generator>`_
-    produced by the JavaScript runtime, but it may be a custom object trying
-    hard to pretend to be a generator. It should have
-    `next <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Generator/next>`_,
-    `return <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Generator/return>`_, and
-    `throw <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Generator/throw>`_ methods.
+    :js:data:`Symbol.toStringTag` is ``"Generator"``. Most likely this will be
+    because it is a true :js:class:`Generator` produced by the JavaScript
+    runtime, but it may be a custom object trying hard to pretend to be a
+    generator. It should have :js:meth:`~Generator.next`,
+    :js:meth:`~Generator.return` and :js:meth:`~Generator.throw` methods.
     """
 
     _js_type_flags = ["IS_GENERATOR"]
@@ -725,24 +742,28 @@ class JsGenerator(JsIterable[Tco], Generic[Tco, Tcontra, Vco]):
     @overload
     def throw(
         self,
-        __typ: type[BaseException],
-        __val: BaseException | object = ...,
-        __tb: TracebackType | None = ...,
+        typ: type[BaseException],
+        val: BaseException | object = ...,
+        tb: TracebackType | None = ...,
+        /,
     ) -> Tco:
         ...
 
     @overload
     def throw(
-        self, __typ: BaseException, __val: None = ..., __tb: TracebackType | None = ...
+        self,
+        typ: BaseException,
+        val: None = ...,
+        tb: TracebackType | None = ...,
+        /,
     ) -> Tco:
         ...
 
+    @docs_argspec("(self, error: BaseException, /) -> Tco")
     def throw(
         self,
-        type,
-        value,
-        traceback,
-    ):
+        *args: Any,
+    ) -> Tco:
         """
         Raises an exception at the point where the generator was paused, and
         returns the next value yielded by the generator function.
@@ -755,7 +776,7 @@ class JsGenerator(JsIterable[Tco], Generic[Tco, Tcontra, Vco]):
         In typical use, this is called with a single exception instance similar
         to the way the raise keyword is used.
 
-        For backwards compatibility, however, the second signature is supported,
+        For backwards compatibility, however, a second signature is supported,
         following a convention from older versions of Python. The type argument
         should be an exception class, and value should be an exception instance.
         If the value is not provided, the type constructor is called to get an
@@ -770,11 +791,11 @@ class JsGenerator(JsIterable[Tco], Generic[Tco, Tcontra, Vco]):
         function was paused.
 
         If the generator function then exits gracefully, is already closed, or
-        raises :any:`GeneratorExit` (by not catching the exception), close
+        raises :any:`GeneratorExit` (by not catching the exception), ``close()``
         returns to its caller. If the generator yields a value, a
         :any:`RuntimeError` is raised. If the generator raises any other
-        exception, it is propagated to the caller. close() does nothing if the
-        generator has already exited due to an exception or normal exit.
+        exception, it is propagated to the caller. ``close()`` does nothing if
+        the generator has already exited due to an exception or normal exit.
         """
 
     def __next__(self) -> Tco:
@@ -785,6 +806,10 @@ class JsGenerator(JsIterable[Tco], Generic[Tco, Tcontra, Vco]):
 
 
 class JsFetchResponse(JsProxy):
+    """A :any:`JsFetchResponse` object represents a :js:data:`Response` to a
+    :js:func:`fetch` request.
+    """
+
     bodyUsed: bool
     ok: bool
     redirected: bool
@@ -807,14 +832,15 @@ class JsFetchResponse(JsProxy):
 
 
 class JsAsyncGenerator(JsAsyncIterable[Tco], Generic[Tco, Tcontra, Vco]):
-    """A JavaScript async generator
+    """A JavaScript :js:class:`AsyncGenerator`
 
     A JavaScript object is treated as an async generator if it's
-    `Symbol.toStringTag <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol/toStringTag>`_
-    is ``"AsyncGenerator"``. Most likely this will be because it is a true async
-    generator produced by the JavaScript runtime, but it may be a custom object
-    trying hard to pretend to be an async generator. It should have ``next``,
-    ``return``, and ``throw`` methods.
+    :js:data:`Symbol.toStringTag` is ``"AsyncGenerator"``. Most likely this will
+    be because it is a true async generator produced by the JavaScript runtime,
+    but it may be a custom object trying hard to pretend to be an async
+    generator. It should have :js:meth:`~AsyncGenerator.next`,
+    :js:meth:`~AsyncGenerator.return`, and :js:meth:`~AsyncGenerator.throw`
+    methods.
     """
 
     _js_type_flags = ["IS_ASYNC_GENERATOR"]
@@ -825,16 +851,16 @@ class JsAsyncGenerator(JsAsyncIterable[Tco], Generic[Tco, Tcontra, Vco]):
     def __aiter__(self) -> "JsAsyncGenerator[Tco, Tcontra, Vco]":
         raise NotImplementedError
 
-    def asend(self, value: Tcontra) -> Awaitable[Tco]:
+    def asend(self, value: Tcontra, /) -> Awaitable[Tco]:
         """Resumes the execution and "sends" a value into the async generator
         function.
 
         The ``value`` argument becomes the result of the current yield
-        expression. The awaitable returned by the asend() method will return the
-        next value yielded by the generator or raises :any:`StopAsyncIteration`
-        if the asynchronous generator returns. If the generator returned a
-        value, this value is discarded (because in Python async generators
-        cannot return a value).
+        expression. The awaitable returned by the ``asend()`` method will return
+        the next value yielded by the generator or raises
+        :any:`StopAsyncIteration` if the asynchronous generator returns. If the
+        generator returned a value, this value is discarded (because in Python
+        async generators cannot return a value).
 
         When ``asend()`` is called to start the generator, the argument will be
         ignored. Unlike in Python, we cannot detect that the generator hasn't
@@ -846,28 +872,29 @@ class JsAsyncGenerator(JsAsyncIterable[Tco], Generic[Tco, Tcontra, Vco]):
     @overload
     def athrow(
         self,
-        __typ: type[BaseException],
-        __val: BaseException | object = ...,
-        __tb: TracebackType | None = ...,
+        typ: type[BaseException],
+        val: BaseException | object = ...,
+        tb: TracebackType | None = ...,
+        /,
     ) -> Awaitable[Tco]:
         ...
 
     @overload
     def athrow(
-        self, __typ: BaseException, __val: None = ..., __tb: TracebackType | None = ...
+        self,
+        typ: BaseException,
+        val: None = ...,
+        tb: TracebackType | None = ...,
+        /,
     ) -> Awaitable[Tco]:
         ...
 
-    def athrow(
-        self,
-        type,
-        value,
-        traceback,
-    ):
+    @docs_argspec("(self, error: BaseException, /) -> Tco")
+    def athrow(self, value: Any, *args: Any) -> Awaitable[Tco]:
         """Resumes the execution and raises an exception at the point where the
         generator was paused.
 
-        The awaitable returned by the asend() method will return the next value
+        The awaitable returned by ``athrow()`` method will return the next value
         yielded by the generator or raises :any:`StopAsyncIteration` if the
         asynchronous generator returns. If the generator returned a value, this
         value is discarded (because in Python async generators cannot return a
@@ -878,15 +905,15 @@ class JsAsyncGenerator(JsAsyncIterable[Tco], Generic[Tco, Tcontra, Vco]):
         raise NotImplementedError
 
     def aclose(self) -> Awaitable[None]:
-        """Raises a :any:`GeneratorExit` at the point where the generator function was
-        paused.
+        """Raises a :any:`GeneratorExit` at the point where the generator
+        function was paused.
 
         If the generator function then exits gracefully, is already closed, or
-        raises :any:`GeneratorExit` (by not catching the exception), close returns to
-        its caller. If the generator yields a value, a :any:`RuntimeError` is raised.
-        If the generator raises any other exception, it is propagated to the
-        caller. close() does nothing if the generator has already exited due to
-        an exception or normal exit.
+        raises :any:`GeneratorExit` (by not catching the exception),
+        ``aclose()`` returns to its caller. If the generator yields a value, a
+        :any:`RuntimeError` is raised. If the generator raises any other
+        exception, it is propagated to the caller. ``aclose()`` does nothing if
+        the generator has already exited due to an exception or normal exit.
         """
         raise NotImplementedError
 
@@ -920,6 +947,8 @@ class JsRawException(JsProxy):
 class JsException(Exception):
     """
     A wrapper around a JavaScript Error to allow it to be thrown in Python.
+
+    The js_error field contains a JsProxy for the wrapped error.
     See :ref:`type-translations-errors`.
     """
 
@@ -931,6 +960,12 @@ class JsException(Exception):
 
 class ConversionError(Exception):
     """An error thrown when conversion between JavaScript and Python fails."""
+
+
+class InternalError(Exception):
+    """Thrown when a recoverable assertion error occurs in internal Pyodide code"""
+
+    pass
 
 
 class JsDomElement(JsProxy):
@@ -956,7 +991,7 @@ class JsDomElement(JsProxy):
 
 
 def create_once_callable(obj: Callable[..., Any], /) -> JsOnceCallable:
-    """Wrap a Python callable in a JavaScript function that can be called once.
+    """Wrap a Python Callable in a JavaScript function that can be called once.
 
     After being called the proxy will decrement the reference count
     of the Callable. The JavaScript function also has a ``destroy`` API that
@@ -1094,7 +1129,8 @@ def to_js(
         can just iterate over the ``Array`` and destroy the proxies).
 
     create_pyproxies:
-        If you set this to False, :any:`to_js` will raise an error
+        If you set this to :any:`False`, :any:`to_js` will raise an error rather
+        than creating any pyproxies.
 
     dict_converter:
         This converter if provided receives a (JavaScript) iterable of
@@ -1119,8 +1155,7 @@ def to_js(
     argument.
 
 
-    In addition to the normal conversions, convert JavaScript `Date
-    <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date>`_
+    In addition to the normal conversions, convert JavaScript :js:class:`Date`
     objects to :py:class:`~datetime.datetime` objects:
 
     .. code-block:: python